+++ src/test/java/com/pivovarit/collectors/inParallelToList/ToListExceptionPropagationTest.java
@@ -1,6 +1,6 @@
 package com.pivovarit.collectors.inParallelToList;
 
-import com.pivovarit.collectors.ExecutorAwareTest;
+import com.pivovarit.collectors.infrastructure.ExecutorAwareTest;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.CompletionException;
