+++ src/test/java/com/pivovarit/collectors/inParallelToSet/ToSetParallelismThrottlingTest.java
@@ -10,7 +10,7 @@ import java.util.stream.Stream;
 
 import static com.pivovarit.collectors.ParallelCollectors.inParallelToSet;
 import static com.pivovarit.collectors.ParallelCollectors.supplier;
-import static com.pivovarit.collectors.TimeUtils.returnWithDelay;
+import static com.pivovarit.collectors.infrastructure.TimeUtils.returnWithDelay;
 import static java.time.Duration.ofMillis;
 import static org.assertj.core.api.Assertions.assertThat;
